moved comparison classes from common to fwd_declare. This allows me to not have to include common.h in any header files
