+++ include/coreir/ir/fwd_declare.h
@@ -57,6 +57,8 @@ class NamedType;
 
 class TypeGen;
 
+
+
 class TypeCache;
 class ValueCache;
 
@@ -111,6 +113,7 @@ class PassManager;
 typedef std::map<std::string,Value*> Values;
 typedef std::map<std::string,ValueType*> Params;
 
+
 bool operator==(const Values& l, const Values& r);
 
 
@@ -128,6 +131,30 @@ typedef std::vector<std::reference_wrapper<const std::string>> ConstSelectPath;
 typedef std::pair<Wireable*,Wireable*> Connection;
 //This is meant to be in relation to an instance. First wireable of the pair is of that instance.
 typedef std::vector<std::pair<Wireable*,Wireable*>> LocalConnections;
+
+
+//Comparison classes for caches
+class ConnectionComp {
+  public:
+    static bool SPComp(const SelectPath& l, const SelectPath& r);
+    bool operator() (const Connection& l, const Connection& r) const;
+};
+
+class ConnectionStrComp {
+  public:
+    static bool SPComp(const SelectPath& l, const SelectPath& r);
+    bool operator() (const Connection& l, const Connection& r) const;
+};
+
+class ValuesComp {
+  public:
+    bool operator() (const Values& l, const Values& r) const;
+};
+
+
+
+
+
 
 //TODO This stuff is super fragile. 
 // Magic hash function I found online
