+++ include/coreir/ir/moduledef.h
@@ -3,7 +3,6 @@
 
 
 #include "fwd_declare.h"
-#include "common.h"
 #include "context.h"
 #include "module.h"
 #include "wireable.h"
