+++ src/ir/generatordef.cpp
@@ -1,5 +1,6 @@
 #include "coreir/ir/generatordef.h"
 #include "coreir/ir/generator.h"
+#include "coreir/ir/common.h"
 
 namespace CoreIR {
 void GeneratorDefFromFun::createModuleDef(ModuleDef* mdef, Values genargs) {
